FRAAS-814 Remove ForgeOps commit SHA label from Docker images (#56)
Removing the recently introduced label for the ForgeOps commit SHA which was last referenced
by forgeops-init. Adding this label may result in us rebuilding one of the product Docker images
when we do not need to - Which would have the knock-on effect of upgrading Identity Stack
Helm charts needlessly and unnecessary downtime while these deployments restart.

We can verify that the set of Docker images we're including in an org-engine build are based on
product Docker images which have undergone PIT testing together without needing this label.

This commit also reduces the set of references which we record when running the script. We
now just store the forgeops commit SHA as all other references can be derived from this.